// src/app/models/book.ts
export interface BookInterface {
  titulo?: string;
  idioma?: string;
  descripcion?: string;
  likes?: any;

  precio?: string;
  link_amazon?: string;
  autor?: string;
  oferta?: string;
  id?: string;
  userUid?: string;
  provincia?: string;
  telefone?: string;
  cidade?: string;
  email?: string;
  situacao?: string;
  star?: string;
  createdAt?: number;
  vagas?: number;
  urlImage?: string;
  urlImageVerso1?: string;
  urlImageVerso2?: string;
  urlImageVerso3?: string;
  Mydate?: string;
  timestamp?: string;
  bairro?: string;
  views?: number;
  categorias?: string;
  phontUrl?: any;
}


export interface Views {
views?: number;
}