// src/app/models/user.ts
// export interface Roles {
//   editor?: boolean;
//   admin?: boolean;
// }

export interface UserInterface {
  id?: string;
  uid?: string;
  name?: string;
  firstName?: string;
  lastName?: string;
  email?: string;
  password?: string;
  createdAt?: any;
  photoUrl?: any;
  userName?: string;
  // roles: Roles;
  location?: string;
  phone?: string;
  phoneNumber?: any;
  displayName?: string;
  bio?: string;
  website?: string;
  customer?: string;
  userUid?: string;
  emailVerified?: any;
  followers?: any;
  following?: any;
  favors?: any;
  brand?: string;
}